+++ src/transaction.ts
@@ -1,4 +1,10 @@
 import { Pool, PoolClient } from 'pg'
+import { DatabaseError } from 'pg-protocol'
+import { isFunction } from './util'
+
+const serializationFailure = '40001'
+const deadlockDetected = '40P01'
+const noActiveSqlTransaction = '25P01'
 
 /**
  * The isolation level to use within a transaction.
@@ -37,87 +43,91 @@ export enum AccessMode {
   ReadOnly = 'READ ONLY',
 }
 
-/**
- * The isolation level and access mode to use within a transaction.
- *
- * See
- * {@link https://www.postgresql.org/docs/current/transaction-iso.html Transaction isolation}
- * in the PostgreSQL manual for more information.
- */
-export interface TransactionMode {
-  isolationLevel: IsolationLevel
-  accessMode: AccessMode
+export interface TransactionOptions {
+  /**
+   * The access mode of the transaction. It may be either:
+   *
+   * - `AccessMode.Default`
+   * - `AccessMode.ReadWrite`
+   * - `AccessMode.ReadOnly`
+   *
+   * See
+   * {@link https://www.postgresql.org/docs/current/sql-set-transaction.html SET TRANSACTION}
+   * in the PostgreSQL manual for more information.
+   */
+  accessMode?: AccessMode
+  /**
+   * The isolation level of the transaction. It may be either:
+   *
+   * - `IsolationLevel.Default`
+   * - `IsolationLevel.Serializable`
+   * - `IsolationLevel.RepeatableRead`
+   * - `IsolationLevel.ReadCommitted`
+   *
+   * See
+   * {@link https://www.postgresql.org/docs/current/transaction-iso.html Transaction isolation}
+   * in the PostgreSQL manual for more information.
+   */
+  isolationLevel?: IsolationLevel
+  /**
+   * The maximum number of times to retry the transaction. Defaults to 2.
+   */
+  maxRetries?: number
+  /**
+   * Whether to retry the transaction in case of an error. By default,
+   * PostgreSQL errors codes `40001` (serialization failure) and `40P01`
+   * (deadlock detected) are considered to be retryable.
+   */
+  shouldRetry?: (err: Error) => boolean
 }
 
-const defaultTransactionMode: TransactionMode = {
-  isolationLevel: IsolationLevel.Default,
-  accessMode: AccessMode.Default,
-}
+const defaultTransactionOptions: TransactionOptions = {}
 
 /**
  * Execute a set of queries within a transaction.
  *
  * Start a transaction and execute a set of queries within it. If the function
- * does not throw an error, the transaction is committed. Returns the value
- * returned from the function.
+ * does not throw an error, the transaction is committed and its return value
+ * is returned.
  *
- * If the function throws any kind of error, the transaction is rolled back and
- * the error is rethrown.
+ * If the function throws a non-retryable error, the transaction is rolled back
+ * and the error is rethrown.
+ *
+ * If the function throws a retryable error, the transaction is rolled back and
+ * retried up to 2 or `maxRetries` times. By default, PostgreSQL errors codes
+ * `40001` (serialization failure) and `40P01` (deadlock detected) are
+ * considered to be retryable, but you may customize the behavior by supplying
+ * a custom `shouldRetry` predicate.
+ *
+ * You may also configure the access mode mode and isolation level of the
+ * transaction by supplying the `accessMode` and `isolationLevel` options,
+ * respectively.
  *
  * @param client A connection pool or a client checked out from a pool.
  * @param queries A set of queries to execute within the transaction.
+ * @param options An optional options object.
  */
-export function withTransaction<T>(
+export async function withTransaction<T>(
   client: Pool | PoolClient,
-  queries: (tx: PoolClient) => PromiseLike<T>
+  queries: (tx: PoolClient) => PromiseLike<T>,
+  options: TransactionOptions = defaultTransactionOptions
 ): Promise<T> {
-  return withTransactionMode(defaultTransactionMode, client, queries)
-}
-
-/**
- * Execute a set of queries within a transaction, using the given isolation
- * level.
- *
- * @param isolationLevel The isolation level to use.
- * @param client A connection pool or a client checked out from a pool.
- * @param queries A set of queries to execute within the transaction.
- */
-export function withTransactionLevel<T>(
-  isolationLevel: IsolationLevel,
-  client: Pool | PoolClient,
-  queries: (tx: PoolClient) => PromiseLike<T>
-): Promise<T> {
-  return withTransactionMode(
-    { ...defaultTransactionMode, isolationLevel },
-    client,
-    queries
+  const beginStatement = getBeginStatement(
+    options.isolationLevel,
+    options.accessMode
   )
-}
-
-/**
- * Execute a set of queries within a transaction, using the given isolation
- * level and access mode.
- *
- * @param transactionMode The isolation level and access mode to use.
- * @param client A connection pool or a client checked out from a pool.
- * @param queries A set of queries to execute within the transaction.
- */
-export async function withTransactionMode<T>(
-  transactionMode: TransactionMode,
-  client: Pool | PoolClient,
-  queries: (tx: PoolClient) => PromiseLike<T>
-): Promise<T> {
-  const beginStatement = getBeginStatement(transactionMode)
+  const shouldRetry = getShouldRetry(options.shouldRetry)
+  const maxRetries = getMaxRetries(options.maxRetries)
   const tx = client instanceof Pool ? await client.connect() : client
 
   try {
-    await tx.query(beginStatement)
-    const result = await queries(tx)
-    await tx.query('COMMIT')
-    return result
-  } catch (err) {
-    await tx.query('ROLLBACK')
-    throw err
+    return await performTransaction(
+      tx,
+      beginStatement,
+      queries,
+      shouldRetry,
+      maxRetries
+    )
   } finally {
     if (client instanceof Pool) {
       tx.release()
@@ -125,16 +135,16 @@ export async function withTransactionMode<T>(
   }
 }
 
-function getBeginStatement(transactionMode: TransactionMode): string {
-  return (
-    'BEGIN' +
-    getIsolationLevel(transactionMode.isolationLevel) +
-    getAccessMode(transactionMode.accessMode)
-  )
+function getBeginStatement(
+  isolationLevel: IsolationLevel | undefined,
+  accessMode: AccessMode | undefined
+): string {
+  return 'BEGIN' + getIsolationLevel(isolationLevel) + getAccessMode(accessMode)
 }
 
-function getIsolationLevel(isolationLevel: IsolationLevel): string {
+function getIsolationLevel(isolationLevel?: IsolationLevel): string {
   switch (isolationLevel) {
+    case undefined:
     case IsolationLevel.Default:
       return ''
     case IsolationLevel.ReadCommitted:
@@ -148,8 +158,9 @@ function getIsolationLevel(isolationLevel: IsolationLevel): string {
   }
 }
 
-function getAccessMode(accessMode: AccessMode): string {
+function getAccessMode(accessMode?: AccessMode): string {
   switch (accessMode) {
+    case undefined:
     case AccessMode.Default:
       return ''
     case AccessMode.ReadWrite:
@@ -158,6 +169,63 @@ function getAccessMode(accessMode: AccessMode): string {
       return ' READ ONLY'
     default:
       throw new TypeError(`Invalid access mode: ${accessMode}`)
+  }
+}
+
+function getShouldRetry(shouldRetry?: (err: Error) => boolean) {
+  if (shouldRetry === undefined) {
+    return isRetryableError
+  } else if (!isFunction(shouldRetry)) {
+    throw new TypeError(`shouldRetry must be a function!`)
+  } else {
+    return shouldRetry
+  }
+}
+
+function getMaxRetries(maxRetries?: number): number {
+  if (maxRetries === undefined) {
+    return 2
+  } else if (!Number.isInteger(maxRetries) || maxRetries < 0) {
+    throw new TypeError(`maxRetries must be a non-negative integer!`)
+  } else {
+    return maxRetries
+  }
+}
+
+async function performTransaction<T>(
+  tx: PoolClient,
+  beginStatement: string,
+  queries: (tx: PoolClient) => PromiseLike<T>,
+  shouldRetry: (err: Error) => boolean,
+  maxRetries: number
+): Promise<T> {
+  try {
+    await tx.query(beginStatement)
+    const result = await queries(tx)
+    await tx.query('COMMIT')
+    return result
+  } catch (err) {
+    await tx.query('ROLLBACK')
+    if (maxRetries > 0 && shouldRetry(err)) {
+      return performTransaction(
+        tx,
+        beginStatement,
+        queries,
+        shouldRetry,
+        maxRetries - 1
+      )
+    } else {
+      throw err
+    }
+  }
+}
+
+function isRetryableError(err: Error) {
+  if (err instanceof DatabaseError) {
+    const code = err.code
+    return code === serializationFailure || code === deadlockDetected
+  } else {
+    return false
   }
 }
 
@@ -187,8 +255,9 @@ export async function withSavepoint<T>(
     await tx.query('RELEASE SAVEPOINT possu_savepoint')
     return result
   } catch (err) {
-    // no_active_sql_transaction https://www.postgresql.org/docs/current/errcodes-appendix.html
-    if (err?.code !== '25P01') {
+    if (
+      !(err instanceof DatabaseError && err.code === noActiveSqlTransaction)
+    ) {
       await tx.query(
         'ROLLBACK TO SAVEPOINT possu_savepoint; RELEASE SAVEPOINT possu_savepoint'
       )