+++ src/util.ts
@@ -2,6 +2,12 @@ export function isString(value: unknown): value is string {
   return typeof value === 'string'
 }
 
+export function isFunction(
+  value: unknown
+): value is (...args: unknown[]) => unknown {
+  return typeof value === 'function'
+}
+
 export function coerce<T>(value: unknown): T {
   return value as T
 }